+++ Loop/Managers/Store Protocols/GlucoseStoreProtocol.swift
@@ -8,8 +8,10 @@
 
 import LoopKit
 import HealthKit
+import LoopAlgorithm
 
 protocol GlucoseStoreProtocol: AnyObject {
+    var latestGlucose: GlucoseSampleValue? { get }
     func getGlucoseSamples(start: Date?, end: Date?) async throws -> [StoredGlucoseSample]
     func addGlucoseSamples(_ samples: [NewGlucoseSample]) async throws -> [StoredGlucoseSample]
 }
